// src/ipc/generated/Conversation.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Message } from "./Message";

export type Conversation = { id: string, characterId: string, title: string, messages: Array<Message>, schemaVersion: number, createdAt: string, updatedAt: string, };